Fix Error in Package Generation
This error arises, nondeterministically, depending on the order in which MDMs are added to the list of MDMs to process. Java’s `File.listContents` method does not promise a particular order in which the contents of a directory will be returned. On many machines, the order that was chosen masked this error: CMASI’s sub-packaged (vehicle, impact, perceive) would be generated before CMASI was generated. In that case, the code worked correctly.

But when CMASI was generated first — with the correct contents of afrl-cmasi.ads — the later generation of vehicle, impact and perceive would overwrite afrl-cmasi.ads, because of the unconditional write. Making the write conditional on the file not existing solves the problem.

+++ src/avtas/lmcp/lmcpgen/AdaMethods.java
@@ -344,8 +344,14 @@ public class AdaMethods {
             packageFileDir = packageFileDir.getParentFile();
             packageFileDir.mkdirs();
             File packageFile = new File(packageFileDir, packageName + ".ads");
-            packageFile.createNewFile();
-            Files.write(packageFile.toPath(), str.getBytes());
+            
+            // We do not want to write the package file if it already exists,
+            // because we may have previously written it, with necessary withs,
+            // on an earlier mdm. (This is particularly true for CMASI MDMs.)
+            if (!packageFile.exists()) {
+            	packageFile.createNewFile();
+            	Files.write(packageFile.toPath(), str.getBytes());
+            }
         }
 
         // Lowest level package definition text, returned as str
